requirements.txt
tensorflow==2.15.0
grpcio
grpcio-tools
protobuf
pillow
numpy
requests
python-dotenv